+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/PaymentMethodRegistry.kt
@@ -25,6 +25,7 @@ import com.stripe.android.lpmfoundations.paymentmethod.definitions.MobilePayDefi
 import com.stripe.android.lpmfoundations.paymentmethod.definitions.MultibancoDefinition
 import com.stripe.android.lpmfoundations.paymentmethod.definitions.OxxoDefinition
 import com.stripe.android.lpmfoundations.paymentmethod.definitions.P24Definition
+import com.stripe.android.lpmfoundations.paymentmethod.definitions.PayNowDefinition
 import com.stripe.android.lpmfoundations.paymentmethod.definitions.PayPalDefinition
 import com.stripe.android.lpmfoundations.paymentmethod.definitions.RevolutPayDefinition
 import com.stripe.android.lpmfoundations.paymentmethod.definitions.SatispayDefinition
@@ -66,6 +67,7 @@ internal object PaymentMethodRegistry {
         MultibancoDefinition,
         OxxoDefinition,
         P24Definition,
+        PayNowDefinition,
         PayPalDefinition,
         RevolutPayDefinition,
         SatispayDefinition,